requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
SQLAlchemy==2.0.32
pandas>=2.2.2
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
email-validator>=2.0.0

